+++ gallery/src/Component/Pixabay.css
@@ -1,7 +1,6 @@
 .apiArea{
   width: 100%;
   background-color: #bbb;
-  /* padding-bottom:10%; */
   height: 440px;
   overflow:hidden ;
   position: relative;
@@ -11,14 +10,13 @@
   position: absolute;
   top: 10%;
   left: 40%;
-
 }
 
 .imgArea{    
   margin-top: 50px;
   height: 900px;
   justify-content: space-around;
-  /* border: 1px solid green; */
+
 }
 
 div.img{
@@ -26,8 +24,8 @@ div.img{
   height: 248px;
   float: left;
   margin: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
   /* border: 1px solid black; */
-
 }
 
 div.img img{
@@ -46,8 +44,6 @@ div.img img{
   /*여기서 이미지영역 넓이 지정*/
   width: 90%;
   height: 2000px;
-  /* border: 1px solid red; */
-
 }
 
 
@@ -58,11 +54,9 @@ div.img img{
   left: 0;
   width: 100%;
   height: auto;
-  /* background-color: white;  */
-  /* border: 1px solid orange; */
   display: flex;
-  justify-content: flex-end;
-  z-index: 2;
+  justify-content: space-between;
+  z-index: 3;
 }
 
 #topMenuBar > *{
@@ -71,13 +65,47 @@ div.img img{
  margin-right: 15px;
 }
 
+
+.logo{
+  left: 10px;
+  line-height: 40px;
+  margin-left: 20px;
+  color: white;
+  font-weight: bold;
+  font-size: 30px;
+}
+
+.scrolled-logo{
+  left: 10px;
+  line-height: 40px;
+  margin-left: 20px;
+  color: black;
+  font-weight: bold;
+  font-size: 30px;
+}
+
+.info{
+  color: white;
+  margin-bottom: 30px;
+}
+
+.title{
+  font-size: 40px;
+  font-weight: bold;
+  margin-bottom: 5px;
+}
+
+.body{
+  font-size: 18px;
+}
+
 .search{
+  height: auto;
   width: 60%;
   position: absolute;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  
 }
 
 .search input{
@@ -85,22 +113,21 @@ div.img img{
   height: auto;
   border: 1px solid #bbb;
   border-radius: 40px;
-  padding: 3% 3%;
+  padding: 20px 20px;
   font-size: 16px;
 }
 
 .top-search{
-  width: 60%;
-  padding-right: 50px;
+  width: 40%;
 }
 
 .top-search input{
   width: 100%;
-  border: 0px solid #bbb;
+  border: 1px solid #bbb;
   border-radius: 24px;
   padding: 10px 12px;
   font-size: 16px;
-  background-color: #dbdbdb;
+  background-color: #ffffff;
 }
 
 .scroll-color {
@@ -111,11 +138,6 @@ div.img img{
 .scrolled-color { 
   background-color:transparent;
   transition: 0.3s;
-}
-
-.banner img{
-  width: 100%;
-  height: 100%;
 }
 
 .selectImg{
@@ -129,7 +151,6 @@ div.img img{
 }
 
 .icons{
-  float: right;
   display: flex;
 }
 
@@ -140,7 +161,7 @@ div.img img{
   left: 0;
   width: 100%;
   height: 100%;
-  background: rgba(0, 0, 0, 0); /* 반투명한 어두운 배경 */
+  background-color: rgba(0, 0, 0, 0.3); /* 반투명한 어두운 배경 */
   z-index: 9998; /* 팝업 미만의 z-index 값 설정 */
 }
 
@@ -178,61 +199,177 @@ background-color: rgb(11, 207, 11);
 border-radius: 16px;
 }
 
+.download-button:hover{
+  font-size: 26px;
+  background-color: rgb(25, 230, 25);
+}
+
 .link{
   line-height: 40px;
   text-decoration-line: none;
   text-align: center;
-  border: 1px solid rgba(173, 173, 173, 0.8);;
   margin-right: 20px;
   width : auto;
   padding-right: 10px;
   padding-left: 10px;
   height : 40px;
+  background-color: rgba(0, 0, 0, 0);
   border-radius: 40px;
-  background-color: rgba(173, 173, 173, 0.8);
-  color: black;
+  color: white;
   font-weight: bold;
   font-size: 14px;
 } 
 
 .link:hover{
-  color: white;
-  background-color: rgba(173, 173, 173, 0.8);
+  background-color: rgba(173, 173, 173, 0.356);
   font-size: 14px;
 }
+
+.scrolled-link{
+  line-height: 40px;
+  text-decoration-line: none;
+  text-align: center;
+  margin-right: 20px;
+  width : auto;
+  padding-right: 10px;
+  padding-left: 10px;
+  height : 40px;
+  background-color: rgba(0, 0, 0, 0);
+  border-radius: 40px;
+  color: black;
+  font-weight: bold;
+  font-size: 14px;
+}
+
+.scrolled-link:hover{
+  background-color: rgba(173, 173, 173, 0.356);
+  font-size: 14px;
+}
+
+
+.more{
+  border: none;
+  margin-right: 20px;
+  width : auto;
+  padding-left: 10px;
+  padding-right: 10px;
+  height : 40px;
+  border-radius: 40px;
+  background-color: transparent;
+  color: white;
+  font-weight: bold;
+  font-size: 14px;
+}
+
+.more:hover{
+  background-color: rgba(173, 173, 173, 0.356);
+  font-size: 14px;
+}
+
+.scrolled-more{
+  border: none;
+  margin-right: 20px;
+  width : auto;
+  padding-left: 10px;
+  padding-right: 10px;
+  height : 40px;
+  border-radius: 40px;
+  background-color: transparent;
+  color: black;
+  font-weight: bold;
+  font-size: 14px;
+}
+
+.scrolled-more:hover{
+  background-color: rgba(173, 173, 173, 0.356);
+  font-size: 14px;
+}
+
+
+
 
 .uploadButton{
   border: none;
   margin-right: 35px;
-  width : auto;
+  width : 90px;
   padding-left: 20px;
   padding-right: 20px;
   height : 40px;
-  border-radius: 40px;
-  background-color: rgb(26, 179, 26);
+  border-radius: 20px;
+  background-color: rgb(4, 168, 92);
   color: rgb(255, 255, 255);
   font-weight: bold;
   font-size: 15px;
 } 
 
 .uploadButton:hover{
-color: white;
-background-color:rgb(29, 192, 29);
-font-size: 15px;
+  color: white;
+  background-color: rgb(10, 192, 107);
+  font-size: 15px;
 }
 
 .UploadPopup {
-border-radius: 10px;
-width: auto;
-height: auto;
-z-index: 9999;
-position: fixed;
-top: 50%;
-left: 50%;
-transform: translate(-50%, -50%);
-background: #fff;
-padding: 20px;
-border: 1px solid rgb(189, 187, 187);
+  border-radius: 10px;
+  width: auto;
+  height: auto;
+  z-index: 9999;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background: #fff;
+  padding: 20px;
+  border: 1px solid rgb(189, 187, 187);
+}
+
+.menuPopup{
+  width: 200px;
+  height: auto;
+  z-index: 9999;
+  position: fixed;
+  top: 60px;
+  right: 160px;
+  background-color: rgb(29, 29, 39);
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  border: 1px solid;
+  border-radius: 10px;
+  overflow: hidden;
 }
 
 
+.menuPopup button{
+  border: none;
+  border-radius:0;
+  color: white;
+  width: 100%;
+  background-color: rgba(255, 255, 255, 0);
+  margin: 0;
+}
+
+.menuPopup button:hover{
+  background-color: rgb(83, 83, 109);
+}
+
+
+.menuPopup .menuBtn{
+  color: white;
+  line-height: 40px;
+  text-decoration-line: none;
+  text-align: center;
+  background-color: rgba(0, 0, 0, 0);
+}
+
+.menuPopup .menuBtn:hover{
+  background-color: rgb(83, 83, 109);
+}
+
+.menuBack{
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 9998; /* 팝업 미만의 z-index 값 설정 */
+}
