+++ gallery/src/Component/Upload.css
@@ -22,7 +22,7 @@
     margin-bottom: 10px;
     width: 80%;
     height: 60px;
-    
+    font-size: 20px;
 }
 
 .preview{
@@ -51,13 +51,19 @@
 }
 
 .upload-button{
-    margin-top: 10px;
+    border: none;
+    font-size: 20px;
+    color: white;
+    margin-top: 20px;
     width: 100%;
     height: 60px;
+    background-color: rgb(4, 168, 92);
+    font-weight: bold;
 }
 
 .upload-button:hover{
-    background-color: gray;
+    background-color: rgb(10, 192, 107);
+    font-size: 22px;
 }
 
 .dragArea{
@@ -71,7 +77,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    margin-top: 25%;
+    margin-top: 30%;
 }
 
 .filebox label {
@@ -94,7 +100,6 @@
   }
 
 
-
 .fileInfo{
     display:flex;
     margin-top: 10px;
@@ -105,4 +110,12 @@
 
 .fileInfo p{
     margin-left: 10px;
+}
+
+.uploadComplete{
+    color: green;
+}
+
+.uploadComplete p{
+    margin-left: 0;
 }
